+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [v6.0.17] - 2019-11-04
+
 * Issue **#1315** : Improved health check for missing API key.
 
 * Issue **#1314** : Fixed NPE thrown when logging caused when viewing docs that can't be found.
@@ -1949,7 +1952,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v6.0.16...6.0
+[Unreleased]: https://github.com/gchq/stroom/compare/v6.0.17...6.0
+[v6.0.17]: https://github.com/gchq/stroom/compare/v6.0.16...v6.0.17
 [v6.0.16]: https://github.com/gchq/stroom/compare/v6.0.15...v6.0.16
 [v6.0.15]: https://github.com/gchq/stroom/compare/v6.0.14...v6.0.15
 [v6.0.14]: https://github.com/gchq/stroom/compare/v6.0.13...v6.0.14